+++ libraries/session/src/main/java/androidx/media3/session/MediaSession.java
@@ -902,12 +902,20 @@ public class MediaSession {
     impl.setSessionPositionUpdateDelayMsOnHandler(updateDelayMs);
   }
 
-  /** Sets the {@linkplain Listener listener}. */
+  /**
+   * Sets the {@linkplain Listener listener}.
+   *
+   * <p>This method must be called on the main thread.
+   */
   /* package */ final void setListener(Listener listener) {
     impl.setMediaSessionListener(listener);
   }
 
-  /** Clears the {@linkplain Listener listener}. */
+  /**
+   * Clears the {@linkplain Listener listener}.
+   *
+   * <p>This method must be called on the main thread.
+   */
   /* package */ final void clearListener() {
     impl.clearMediaSessionListener();
   }
@@ -1439,7 +1447,11 @@ public class MediaSession {
     default void onRenderedFirstFrame(int seq) throws RemoteException {}
   }
 
-  /** Listener for media session events */
+  /**
+   * Listener for media session events.
+   *
+   * <p>All methods must be called on the main thread.
+   */
   /* package */ interface Listener {
 
     /**